requirements.txt
aniso8601==9.0.1
arrow==1.2.3
asgiref==3.6.0
autopep8==2.0.2
dateparser==1.1.7
Django==4.1.7
django-appconf==1.0.5
django-filter==22.1
django-jwt-auth-middleware==0.1.1
Faker==12.0.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
mixer==7.2.2
moment==0.12.1
promise==2.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
times==0.7
tzdata==2022.7
tzlocal==4.2
